ci: remove cxx11 suffixes from benchmarks
