+++ CMakeLists.txt
@@ -35,7 +35,7 @@ target_include_directories(wide_integer INTERFACE
     ${CMAKE_CURRENT_SOURCE_DIR}/include
 )
 
-target_compile_features(wide_integer INTERFACE cxx_std_17)
+target_compile_features(wide_integer INTERFACE cxx_std_11)
 
 if(WI_BUILD_TESTS)
     find_package(GTest REQUIRED)
@@ -53,77 +53,50 @@ if(WI_BUILD_TESTS)
         gtest_discover_tests(${name} NO_PRETTY_VALUES)
     endfunction()
 
-    add_wide_test(wide_integer_test tests/wide_integer_test.cpp 17)
+    add_wide_test(wide_integer_test tests/wide_integer_test.cpp 11)
     target_link_libraries(wide_integer_test PRIVATE fmt::fmt)
-    add_wide_test(wide_integer_cxx11_test tests/wide_integer_test.cpp 11)
-    target_compile_definitions(wide_integer_cxx11_test PRIVATE USE_CXX11_HEADER)
-    target_link_libraries(wide_integer_cxx11_test PRIVATE fmt::fmt)
 endif()
 
 if(WI_BUILD_BENCHMARKS)
     find_package(benchmark REQUIRED)
 
-    add_executable(perf_cxx17
+    add_executable(perf
         bench/performance.cpp
     )
-    target_include_directories(perf_cxx17 PRIVATE
+    target_include_directories(perf PRIVATE
         ${CMAKE_CURRENT_SOURCE_DIR}/include
     )
-    target_compile_features(perf_cxx17 PRIVATE cxx_std_17)
-    target_link_libraries(perf_cxx17 PRIVATE fmt::fmt benchmark::benchmark)
-    target_compile_options(perf_cxx17 PRIVATE -O3 -DNDEBUG)
-
-    add_executable(perf_cxx11
-        bench/performance.cpp
-    )
-    target_compile_definitions(perf_cxx11 PRIVATE USE_CXX11_HEADER)
-    target_include_directories(perf_cxx11 PRIVATE
-        ${CMAKE_CURRENT_SOURCE_DIR}/include
-    )
-    target_compile_features(perf_cxx11 PRIVATE cxx_std_11)
-    target_link_libraries(perf_cxx11 PRIVATE fmt::fmt benchmark::benchmark)
-    target_compile_options(perf_cxx11 PRIVATE -O3 -DNDEBUG)
+    target_compile_features(perf PRIVATE cxx_std_11)
+    target_link_libraries(perf PRIVATE fmt::fmt benchmark::benchmark)
+    target_compile_options(perf PRIVATE -O3 -DNDEBUG)
 
     add_executable(perf_compare_int128
         bench/compare_int128.cpp
     )
-    target_compile_features(perf_compare_int128 PRIVATE cxx_std_17)
+    target_compile_features(perf_compare_int128 PRIVATE cxx_std_11)
     target_include_directories(perf_compare_int128 PRIVATE
         ${CMAKE_CURRENT_SOURCE_DIR}/include
     )
-    target_link_libraries(perf_compare_int128 PRIVATE wide_integer fmt::fmt benchmark::benchmark)
+    target_link_libraries(perf_compare_int128 PRIVATE fmt::fmt benchmark::benchmark)
     target_compile_options(perf_compare_int128 PRIVATE -O3 -DNDEBUG)
 
-    add_executable(perf_compare_int128_cxx11
-        bench/compare_int128.cpp
-    )
-    target_compile_definitions(perf_compare_int128_cxx11 PRIVATE USE_CXX11_HEADER)
-    target_compile_features(perf_compare_int128_cxx11 PRIVATE cxx_std_11)
-    target_include_directories(perf_compare_int128_cxx11 PRIVATE
-        ${CMAKE_CURRENT_SOURCE_DIR}/include
-    )
-    target_link_libraries(perf_compare_int128_cxx11 PRIVATE fmt::fmt benchmark::benchmark)
-    target_compile_options(perf_compare_int128_cxx11 PRIVATE -O3 -DNDEBUG)
-
-    add_executable(perf_compare_int256_cxx11
+    add_executable(perf_compare_int256
         bench/compare_int256.cpp
     )
-    target_compile_definitions(perf_compare_int256_cxx11 PRIVATE USE_CXX11_HEADER)
-    target_compile_features(perf_compare_int256_cxx11 PRIVATE cxx_std_11)
-    target_include_directories(perf_compare_int256_cxx11 PRIVATE
+    target_compile_features(perf_compare_int256 PRIVATE cxx_std_11)
+    target_include_directories(perf_compare_int256 PRIVATE
         ${CMAKE_CURRENT_SOURCE_DIR}/include
     )
-    target_link_libraries(perf_compare_int256_cxx11 PRIVATE fmt::fmt benchmark::benchmark)
-    target_compile_options(perf_compare_int256_cxx11 PRIVATE -O3 -DNDEBUG)
+    target_link_libraries(perf_compare_int256 PRIVATE fmt::fmt benchmark::benchmark)
+    target_compile_options(perf_compare_int256 PRIVATE -O3 -DNDEBUG)
 
-    add_executable(perf_compare_int256_random_cxx11
+    add_executable(perf_compare_int256_random
         bench/compare_int256_random.cpp
     )
-    target_compile_definitions(perf_compare_int256_random_cxx11 PRIVATE USE_CXX11_HEADER)
-    target_compile_features(perf_compare_int256_random_cxx11 PRIVATE cxx_std_11)
-    target_include_directories(perf_compare_int256_random_cxx11 PRIVATE
+    target_compile_features(perf_compare_int256_random PRIVATE cxx_std_11)
+    target_include_directories(perf_compare_int256_random PRIVATE
         ${CMAKE_CURRENT_SOURCE_DIR}/include
     )
-    target_link_libraries(perf_compare_int256_random_cxx11 PRIVATE fmt::fmt benchmark::benchmark)
-    target_compile_options(perf_compare_int256_random_cxx11 PRIVATE -O3 -DNDEBUG)
+    target_link_libraries(perf_compare_int256_random PRIVATE fmt::fmt benchmark::benchmark)
+    target_compile_options(perf_compare_int256_random PRIVATE -O3 -DNDEBUG)
 endif()
