+++ bench/compare_int128.cpp
@@ -1,16 +1,10 @@
 #include <benchmark/benchmark.h>
 #include <boost/multiprecision/cpp_int.hpp>
 
-#ifdef USE_CXX11_HEADER
-#    include <wide_integer/wide_integer_cxx11.h>
-#else
-#    include <wide_integer/wide_integer.h>
-#endif
+#include <wide_integer/wide_integer.h>
 
 using WInt128 = wide::integer<128, unsigned>;
-using WInt256 = wide::integer<256, unsigned>;
 using BInt128 = boost::multiprecision::uint128_t;
-using BInt256 = boost::multiprecision::uint256_t;
 
 static void BM_Wide128Addition(benchmark::State & state)
 {
@@ -192,126 +186,6 @@ static void BM_Boost128Division(benchmark::State & state)
     }
 }
 
-static void BM_Wide256Addition(benchmark::State & state)
-{
-    WInt256 a = 123456789;
-    WInt256 b = 987654321;
-    for (auto _ : state)
-    {
-        WInt256 c = a;
-        for (int i = 0; i < 1000; ++i)
-        {
-            c += b;
-            benchmark::DoNotOptimize(c);
-        }
-    }
-}
-
-static void BM_Wide256Subtraction(benchmark::State & state)
-{
-    WInt256 a = 1234567890;
-    WInt256 b = 987654321;
-    for (auto _ : state)
-    {
-        WInt256 c = a;
-        for (int i = 0; i < 1000; ++i)
-        {
-            c -= b;
-            benchmark::DoNotOptimize(c);
-        }
-    }
-}
-
-static void BM_Boost256Addition(benchmark::State & state)
-{
-    BInt256 a = 123456789;
-    BInt256 b = 987654321;
-    for (auto _ : state)
-    {
-        BInt256 c = a;
-        for (int i = 0; i < 1000; ++i)
-        {
-            c += b;
-            benchmark::DoNotOptimize(c);
-        }
-    }
-}
-
-static void BM_Boost256Subtraction(benchmark::State & state)
-{
-    BInt256 a = 1234567890;
-    BInt256 b = 987654321;
-    for (auto _ : state)
-    {
-        BInt256 c = a;
-        for (int i = 0; i < 1000; ++i)
-        {
-            c -= b;
-            benchmark::DoNotOptimize(c);
-        }
-    }
-}
-
-static void BM_Wide256Multiplication(benchmark::State & state)
-{
-    WInt256 a = 123456789;
-    WInt256 b = 987654321;
-    for (auto _ : state)
-    {
-        WInt256 c = a;
-        for (int i = 0; i < 100; ++i)
-        {
-            c *= b;
-            benchmark::DoNotOptimize(c);
-        }
-    }
-}
-
-static void BM_Wide256Division(benchmark::State & state)
-{
-    WInt256 a = WInt256{1234567890123456789ULL} << 128;
-    WInt256 b = 987654321;
-    for (auto _ : state)
-    {
-        WInt256 c = a;
-        for (int i = 0; i < 100; ++i)
-        {
-            c /= b;
-            benchmark::DoNotOptimize(c);
-        }
-    }
-}
-
-static void BM_Boost256Multiplication(benchmark::State & state)
-{
-    BInt256 a = 123456789;
-    BInt256 b = 987654321;
-    for (auto _ : state)
-    {
-        BInt256 c = a;
-        for (int i = 0; i < 100; ++i)
-        {
-            c *= b;
-            benchmark::DoNotOptimize(c);
-        }
-    }
-}
-
-static void BM_Boost256Division(benchmark::State & state)
-{
-    BInt256 a = BInt256{1234567890123456789ULL} << 128;
-    BInt256 b = 987654321;
-    for (auto _ : state)
-    {
-        BInt256 c = a;
-        for (int i = 0; i < 100; ++i)
-        {
-            c /= b;
-            benchmark::DoNotOptimize(c);
-        }
-    }
-}
-
 BENCHMARK(BM_Wide128Addition);
 BENCHMARK(BM_Wide128Subtraction);
 BENCHMARK(BM_Builtin128Addition);
@@ -324,13 +198,5 @@ BENCHMARK(BM_Boost128Addition);
 BENCHMARK(BM_Boost128Subtraction);
 BENCHMARK(BM_Boost128Multiplication);
 BENCHMARK(BM_Boost128Division);
-BENCHMARK(BM_Wide256Addition);
-BENCHMARK(BM_Wide256Subtraction);
-BENCHMARK(BM_Boost256Addition);
-BENCHMARK(BM_Boost256Subtraction);
-BENCHMARK(BM_Wide256Multiplication);
-BENCHMARK(BM_Wide256Division);
-BENCHMARK(BM_Boost256Multiplication);
-BENCHMARK(BM_Boost256Division);
 
 BENCHMARK_MAIN();
