+++ bench/compare_int256.cpp
@@ -1,11 +1,7 @@
 #include <benchmark/benchmark.h>
 #include <boost/multiprecision/cpp_int.hpp>
 
-#ifdef USE_CXX11_HEADER
-#    include <wide_integer/wide_integer_cxx11.h>
-#else
-#    include <wide_integer/wide_integer.h>
-#endif
+#include <wide_integer/wide_integer.h>
 
 using WInt = wide::integer<256, signed>;
 using BInt = boost::multiprecision::int256_t;
